requirements.txt
Flask
RPi.GPIO
PyYAML
pytest
requests
